+++ packages/components/pie-button/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 1.14.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-spinner@1.5.4
+  - @justeattakeaway/pie-webc-core@17.0.0
+
+## 1.14.10
+
+### Patch Changes
+
+- Updated dependencies [[`e9ef549`](https://github.com/justeattakeaway/pie/commit/e9ef54989cea1df4f77b9bb2923390318d00fa5e)]:
+  - @justeattakeaway/pie-webc-core@16.1.1
+  - @justeattakeaway/pie-spinner@1.5.3
+
 ## 1.14.9
 
 ### Patch Changes