+++ packages/components/pie-icon-button/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @justeattakeaway/pie-icon-button
 
+## 2.7.28
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-spinner@1.5.4
+  - @justeattakeaway/pie-webc-core@17.0.0
+  - @justeattakeaway/pie-icons-webc@1.27.3
+
+## 2.7.27
+
+### Patch Changes
+
+- Updated dependencies [[`e9ef549`](https://github.com/justeattakeaway/pie/commit/e9ef54989cea1df4f77b9bb2923390318d00fa5e)]:
+  - @justeattakeaway/pie-webc-core@16.1.1
+  - @justeattakeaway/pie-spinner@1.5.3
+  - @justeattakeaway/pie-icons-webc@1.27.2
+
 ## 2.7.26
 
 ### Patch Changes