+++ packages/components/pie-textarea/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @justeattakeaway/pie-textarea
 
+## 0.23.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-assistive-text@1.0.8
+  - @justeattakeaway/pie-webc-core@17.0.0
+
+## 0.23.2
+
+### Patch Changes
+
+- Updated dependencies [[`e9ef549`](https://github.com/justeattakeaway/pie/commit/e9ef54989cea1df4f77b9bb2923390318d00fa5e)]:
+  - @justeattakeaway/pie-webc-core@16.1.1
+  - @justeattakeaway/pie-assistive-text@1.0.7
+
 ## 0.23.1
 
 ### Patch Changes